requirements.txt
flask==1.1.4
Flask-Script==2.0.6
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
mysqlclient==2.0.3
redis==2.10.6
unittest2==1.1.0
requests==2.25.1
email-validator==1.1.2
markupsafe==2.0.1
Jinja2==2.11.3
